refactor : 스타일 분리 및 오타 수정
