+++ src/pages/layout/Layout.styles.ts
@@ -1,0 +1,9 @@
+import styled from '@emotion/styled';
+
+export const LayoutContainer = styled.div`
+  width: 100%;
+  max-width: var(--max-width-container);
+  display: flex;
+  flex-direction: column;
+  gap: 36px;
+`;
