+++ src/pages/layout/Layout.tsx
@@ -1,12 +1,12 @@
-import styled from '@emotion/styled';
+import * as S from './Layout.styles';
 import { Outlet } from 'react-router-dom';
-import Header from '../components/Header';
-import { CartProvider } from '../context/CartContext';
+import Header from '../../components/header/Header';
+import { CartProvider } from '../../context/CartContext';
 import { useEffect, useState } from 'react';
-import tryApiCall from '../util/tryApiCall';
-import { getCartItems } from '../services/cartItemServices';
-import { CartItemType } from '../types/data';
-import { ErrorMessageProvider } from '../context/ErrorMessageContext';
+import tryApiCall from '../../util/tryApiCall';
+import { getCartItems } from '../../services/cartItemServices';
+import type { CartItemType } from '../../types/data';
+import { ErrorMessageProvider } from '../../context/ErrorMessageContext';
 
 const Layout = () => {
   const [cartItemsIds, setCartItemsIds] = useState<number[]>([]);
@@ -38,21 +38,13 @@ const Layout = () => {
         handleAddCartItemsIds={handleAddCartItemsIds}
         handleRemoveCartItemsIds={handleRemoveCartItemsIds}
       >
-        <LayoutContainer>
+        <S.LayoutContainer>
           <Header />
           <Outlet />
-        </LayoutContainer>
+        </S.LayoutContainer>
       </CartProvider>
     </ErrorMessageProvider>
   );
 };
 
 export default Layout;
-
-const LayoutContainer = styled.div`
-  width: 100%;
-  max-width: var(--max-width-container);
-  display: flex;
-  flex-direction: column;
-  gap: 36px;
-`;
