+++ src/types/styleVariants.ts
@@ -1,0 +1,1 @@
+export type VariantsProps = 'smallBlack' | 'smallGrey';
